# docs/swagger.yaml
definitions:
  entity.Words:
    properties:
      id:
        type: string
      key:
        type: string
      value:
        type: integer
    type: object
  server.InsertWordRequest:
    properties:
      words:
        additionalProperties:
          type: integer
        type: object
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  title: Postgres Crud API
  version: "1.0"
paths:
  /new-word:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.InsertWordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Take Word
      tags:
      - exam
  /take-all-word:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Words'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Give All Word
      tags:
      - exam
  /take-limited-word:
    get:
      consumes:
      - application/json
      parameters:
      - description: GET
        in: query
        name: page
        required: true
        type: integer
      - description: GET
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Words'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Give All Word
      tags:
      - exam
swagger: "2.0"